requirements.txt
asgiref==3.5.1
Django==3.2
djangorestframework==3.12.4
importlib-metadata==4.11.3
Markdown==3.3.7
psycopg2==2.9.3
pypng==0.0.21
PyQRCode==1.2.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
zipp==3.8.0
